Generate front and back faces
